+++ electron-builder.config.ts
@@ -66,10 +66,7 @@ const config: Configuration = {
     hardenedRuntime: true,
     entitlements: 'build/entitlements.mac.plist',
     entitlementsInherit: 'build/entitlements.mac.plist',
-    target: [
-      { target: 'dmg', arch: ['arm64'] },
-      { target: 'zip', arch: ['arm64'] },
-    ],
+    target: [{ target: 'dmg', arch: ['arm64'] }],
     icon: 'src/assets/images/yoda/yoda.icns',
     notarize: false,
     extendInfo: {